encoding: Move Frequency from maccommands to types module
