+++ lorawan-encoding/README.md
@@ -24,8 +24,9 @@ lorawan = "0.9"
 ### Packet generation
 
 ```rust
-use lorawan::{creator::JoinAcceptCreator, keys, maccommands};
+use lorawan::{creator::JoinAcceptCreator, keys};
 use lorawan::default_crypto::DefaultFactory;
+use lorawan::types::Frequency;
 use heapless;
 
 fn main() {
@@ -38,9 +39,9 @@ fn main() {
     phy.set_dev_addr(&[1; 4]);
     phy.set_dl_settings(2);
     phy.set_rx_delay(1);
-    let mut freqs: heapless::Vec<lorawan::maccommands::Frequency, 2> = heapless::Vec::new();
-    freqs.push(maccommands::Frequency::new(&[0x58, 0x6e, 0x84,]).unwrap()).unwrap();
-    freqs.push(maccommands::Frequency::new(&[0x88, 0x66, 0x84,]).unwrap()).unwrap();
+    let mut freqs: heapless::Vec<Frequency, 2> = heapless::Vec::new();
+    freqs.push(Frequency::new(&[0x58, 0x6e, 0x84,]).unwrap()).unwrap();
+    freqs.push(Frequency::new(&[0x88, 0x66, 0x84,]).unwrap()).unwrap();
     phy.set_c_f_list(freqs).unwrap();
     let payload = phy.build(&key).unwrap();
     println!("Payload: {:x?}", payload);
