+++ lorawan-encoding/tests/maccommands.rs
@@ -1,5 +1,6 @@
 use lorawan::maccommandcreator::*;
 use lorawan::maccommands::*;
+use lorawan::types::Frequency;
 
 macro_rules! test_helper {
     ( $cmd:ident, $data:ident, $name:ident, $type:ident, $size:expr, $( ( $method:ident, $val:expr ) ,)*) => {{
